requirements.txt
pypng
Pillow
psutil
dill
numpy
scipy
matplotlib
tensorflow==1.15
onnx
pycddlib
